Exception handling modified

+++ src/main/java/com/ams/service/UserServiceImpl.java
@@ -12,6 +12,7 @@ import org.springframework.stereotype.Service;
 import com.ams.db.entity.Role;
 import com.ams.db.entity.User;
 import com.ams.dto.UserDTO;
+import com.ams.exception.AmsException;
 import com.ams.repository.RoleRepository;
 import com.ams.repository.UserRepository;
 
@@ -28,10 +29,15 @@ public class UserServiceImpl implements UserService {
 	private RoleRepository roleDAO;
 
 	@Override
-	public UserDTO getUser(long userId) throws Exception {
+	public UserDTO getUser(long userId) {
 		LOG.debug("getUser() using id  {}", userId);
 		Optional<User> optionalUsers = userDAO.findById(userId);
-		optionalUsers.orElseThrow(() -> new Exception("User not found"));
+		//optionalUsers.orElseThrow(() -> new Exception("User not found"));
+		if(!optionalUsers.isPresent())
+		{
+			throw new AmsException("User not found");
+		}
+		
 		User user = optionalUsers.get();
 		Set<String> roles = null;
 		for (Role role : user.getRoles()) {
@@ -48,7 +54,7 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public long addUser(UserDTO userDto) throws Exception {
+	public long addUser(UserDTO userDto) {
 		Set<Role> roleSet = null;
 		if (userDto.getRoles() != null) {
 			roleSet = new HashSet<>();
@@ -57,7 +63,7 @@ public class UserServiceImpl implements UserService {
 				roleSet.add(roleEntity);
 			}
 		} else {
-			throw new Exception("Role is missing");
+			throw new AmsException("Role is missing");
 		}
 
 		User user = new User(userDto.getLoginName(), userDto.getPassword(), userDto.getPin(), userDto.getGivenName(),
@@ -70,10 +76,15 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public long updateUser(UserDTO userDto) throws Exception {
+	public long updateUser(UserDTO userDto) {
 
 		Optional<User> optionalUsers = userDAO.findById(userDto.getId());
-		optionalUsers.orElseThrow(() -> new Exception("Username not found"));
+	//	optionalUsers.orElseThrow(() -> new Exception("Username not found"));
+		
+		if(!optionalUsers.isPresent())
+		{
+			throw new AmsException("User not found");
+		}
 
 		User userToUpdate = optionalUsers.get();
 
@@ -120,10 +131,14 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public long deleteUser(String loginName) throws Exception {
+	public long deleteUser(String loginName)  {
 
 		Optional<User> optionalUsers = userDAO.findByLoginName(loginName);
-		optionalUsers.orElseThrow(() -> new Exception("User not found"));
+		//optionalUsers.orElseThrow(() -> new Exception("User not found"));
+		if(!optionalUsers.isPresent()) {
+			throw new AmsException("User not found");
+		}
+		
 		User user = optionalUsers.get();
 		long id = user.getGennPID();
 		userDAO.delete(user);
@@ -131,10 +146,13 @@ public class UserServiceImpl implements UserService {
 	}
 
 	@Override
-	public UserDTO getUser(String loginName) throws Exception {
+	public UserDTO getUser(String loginName) {
 		LOG.debug("getUser() using login name  {}", loginName);
 		Optional<User> optionalUsers = userDAO.findByLoginName(loginName);
-		optionalUsers.orElseThrow(() -> new Exception("User not found"));
+		//optionalUsers.orElseThrow(() -> new Exception("User not found"));
+		if(!optionalUsers.isPresent()) {
+			throw new AmsException("User not found");
+		}
 		User user = optionalUsers.get();
 		Set<String> roles = null;
 		for (Role role : user.getRoles()) {
